Migrate which action to new pattern
