+++ lib/aruba/api_actions.rb
@@ -1,0 +1,6 @@
+# Aruba
+module Aruba
+  # Api Actions
+  module ApiActions
+  end
+end
